Added in memory tests to cover the MvcSample.
This change adds some functional tests to cover the MvcSample in the
samples folder. The tests validate that the Mvc pipeline is correctly
integrated with the rest of the components like Hosting and Dependency
Injection and that no end to end scenario is regressed by a change.

The only thing we mock in the execution of the tests is the application
base path in order to make it match the application base path that the
sample would have if it was being executed in a regular context.

For testing views our approach is to use the HtmlAgility pack and check
some elements of the returned response in order to ensure that the
elements on the view get rendered properly and don't fail silently.