Reduce use of TypeActivation for ModelBinders
These model binders don't have any per-request state so there is no need
for them to be type activated. In one case the model binder actually does
its own caching which is being defeated by the fact that we register it as
type activated.

More changes to come in this area.