Making StringCollectionTextWriter use UnicodeEncoding by default.
The Encoding property in StringCollectionTextWriter is not used anywhere. The only reason it is present is because TextWriter forces it to. So introducing a default constructor to avoid passing static encoding in different places in the code.

Removing code which passes UTF-8 because .net uses UTF-16 representation for strings (StringWriter does the same - https://msdn.microsoft.com/en-us/library/system.io.stringwriter.encoding.aspx).