Rewrite of Invoker
This change rewrites the invoker state machine to avoid allocating
_anything_ unless necessary to call into a filter. We also avoid calling
into any `async` method unless necessary to avoid the state machine code
generated by the compiler.

This results in some code duplication, but the improvements are
significant and realized any time that you don't have a particular type of
filter.