[Design] Add 'profiles' for ApiDescription
This PR adds infrastructure for 'profiles' that can be matched against
ApiDescriptions. Once a profile matches an action/description, it can
modify the description to fit the application's common practices and
conventions.

The idea is that each 'profile' corresponds to a kind of logical
operation (look at the examples) like 'get-by-id' - and will update the
ApiDescription to include the kind of things this operation would
normally need to declare.

This should make it easier to customize if you have your own set of
opinionated behaviors, such as using 202-accepted for 'create' operations.

This is a design PR, thus I'm looking for feedback on the design and
layering of the approach. (yes, I know about the functional test
failures, I haven't updated those yet).