Add functional tests for Razor re-compilation
This adds functional tests that show the desired behavior for Razor
re-compilation.

Using the precompilation tool show result in views/pages that don't
support updates.

Using the Rzc (msbuild) will result in views/pages that are only used
if they pass a checksum test on startup.

The rzc support is about to land, but I thought it made sense to spec
out the tests first.